% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{get_hhts}
\alias{get_hhts}
\title{HHTS data retrieval}
\usage{
get_hhts(dyear, level, vars)
}
\arguments{
\item{dyear}{data year or combined data year pattern (e.g. 2017_2019)}

\item{level}{either "h" (household), "p" (person), "d" (day), "t" (trip), or "v" (vehicle)}

\item{vars}{character vector with requested variables}
}
\value{
dataframe with variables and necessary weights
}
\description{
Gets requested Household Travel Survey variables
}
