% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_hts.R
\name{get_psrc_hts}
\alias{get_psrc_hts}
\title{Retrieve PSRC travel survey data}
\usage{
get_psrc_hts(survey_years = c(2017, 2019, 2021, 2023), survey_vars)
}
\arguments{
\item{survey_years}{vector with values that may include 2017, 2019, 2021, 2023}

\item{survey_vars}{desired variables (in addition to id fields, which are returned by default)}
}
\value{
surveyDataTools data object, i.e. named list
}
\description{
Retrieve PSRC travel survey data
}
\author{
Michael Jensen
}
