% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_sss.R
\name{get_psrc_sss}
\alias{get_psrc_sss}
\title{Retrieve PSRC transportation safety & security survey data}
\usage{
get_psrc_sss(survey_vars = NULL)
}
\arguments{
\item{survey_vars}{desired variables (in addition to person and household ids plus `person_weight`, which are returned by default)}
}
\value{
data.table of queried safety survey responses
}
\description{
Retrieve PSRC transportation safety & security survey data
}
\author{
Michael Jensen
}
