% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{hhts_connect}
\alias{hhts_connect}
\title{Connection to survey datasource}
\usage{
hhts_connect(connection = NULL)
}
\arguments{
\item{connection}{existing connection if relevant}
}
\value{
database connection
}
\description{
Connection to survey datasource
}
