% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{hhts_recode_na}
\alias{hhts_recode_na}
\title{HHTS recode missing as NA}
\usage{
hhts_recode_na(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with missing values recoded NA
}
\description{
Recode missing values to NA
}
