% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{hhts_varsearch}
\alias{hhts_varsearch}
\title{Search HHTS variable definitions}
\usage{
hhts_varsearch(regex)
}
\arguments{
\item{regex}{search term}
}
\value{
data.table of filtered variable attributes
}
\description{
Look for a variable name using a search term
}
