% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_age}
\alias{hts_bin_age}
\title{Add a binned age variable
Requires `age` variable}
\usage{
hts_bin_age(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with an additional binned age variable
}
\description{
Add a binned age variable
Requires `age` variable
}
\author{
Michael Jensen
}
