% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-groupings.R
\name{hts_bin_commute_freq}
\alias{hts_bin_commute_freq}
\title{Add simplified commute frequency variable
Requires `commute_freq` variable}
\usage{
hts_bin_commute_freq(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified commute frequency variable
}
\description{
Add simplified commute frequency variable
Requires `commute_freq` variable
}
\author{
Michael Jensen
}
