% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_commute_mode}
\alias{hts_bin_commute_mode}
\title{Add simplified commute mode variable
Requires `commute_mode` variable}
\usage{
hts_bin_commute_mode(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified commute mode variable
}
\description{
Add simplified commute mode variable
Requires `commute_mode` variable
}
\author{
Michael Jensen
}
