% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-trip-vars.R
\name{hts_bin_dest_purpose}
\alias{hts_bin_dest_purpose}
\title{Add destination purpose classifications
Requires `dest_purpose` variable}
\usage{
hts_bin_dest_purpose(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified destination purpose variable
}
\description{
Add destination purpose classifications
Requires `dest_purpose` variable
}
\author{
Michael Jensen
}
