% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-groupings.R
\name{hts_bin_edu}
\alias{hts_bin_edu}
\title{Add simplified education variable
Requires `education` variable}
\usage{
hts_bin_edu(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified education variable
}
\description{
Add simplified education variable
Requires `education` variable
}
\author{
Michael Jensen
}
