% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_gender}
\alias{hts_bin_gender}
\title{Add simplified gender variable
Requires `gender` variable}
\usage{
hts_bin_gender(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified gender variable
}
\description{
Add simplified gender variable
Requires `gender` variable
}
\author{
Michael Jensen
}
