% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-hh-vars.R,
%   R/psrc-hts-recode-person-vars.R
\name{hts_bin_hhsize}
\alias{hts_bin_hhsize}
\title{Add simplified household size variable
Requires `hhsize` variable}
\usage{
hts_bin_hhsize(hts_data)

hts_bin_hhsize(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified household size variable

hts_data with a simplified household size variable
}
\description{
Add simplified household size variable
Requires `hhsize` variable

Add simplified household size variable
Requires `hhsize` variable
}
\author{
Michael Jensen
}
