% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-hh-vars.R
\name{hts_bin_income}
\alias{hts_bin_income}
\title{Add binned income variable
Requires `hhincome_broad` variable}
\usage{
hts_bin_income(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a binned income variable
}
\description{
Add binned income variable
Requires `hhincome_broad` variable
}
\author{
Michael Jensen
}
