% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_lum_sector}
\alias{hts_bin_lum_sector}
\title{Add land use modeling industry classification
Requires `industry` variable}
\usage{
hts_bin_lum_sector(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a land use modeling industry classification variable
}
\description{
Add land use modeling industry classification
Requires `industry` variable
}
\author{
Michael Jensen
}
