% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-trip-vars.R
\name{hts_bin_mode}
\alias{hts_bin_mode}
\title{Add simplified mode classification
Requires `mode_characterization` variable}
\usage{
hts_bin_mode(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified mode variable
}
\description{
Add simplified mode classification
Requires `mode_characterization` variable
}
\author{
Michael Jensen
}
