% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-hh-vars.R
\name{hts_bin_rent_own}
\alias{hts_bin_rent_own}
\title{Add simplified tenure variable
Requires `rent_own` variable}
\usage{
hts_bin_rent_own(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified tenure variable
}
\description{
Add simplified tenure variable
Requires `rent_own` variable
}
\author{
Michael Jensen
}
