% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_sexuality}
\alias{hts_bin_sexuality}
\title{Add simplified sexuality variable
Requires `sexuality` variable}
\usage{
hts_bin_sexuality(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified sexuality variable
}
\description{
Add simplified sexuality variable
Requires `sexuality` variable
}
\author{
Michael Jensen
}
