% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_telecommute_freq}
\alias{hts_bin_telecommute_freq}
\title{Add simplified telework frequency variable
Requires `telecommute_freq` variable}
\usage{
hts_bin_telecommute_freq(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified telework frequency variable
}
\description{
Add simplified telework frequency variable
Requires `telecommute_freq` variable
}
\author{
Michael Jensen
}
