% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_telecommute_trichotomy}
\alias{hts_bin_telecommute_trichotomy}
\title{Add telecommute status variable
Requires `telecommute_freq` and `workplace` variables}
\usage{
hts_bin_telecommute_trichotomy(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with telecommute status variable
}
\description{
Add telecommute status variable
Requires `telecommute_freq` and `workplace` variables
}
\author{
Michael Jensen
}
