% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-groupings.R
\name{hts_bin_telework_time}
\alias{hts_bin_telework_time}
\title{Add simplified telework time variable
Requires `telework_time` variable}
\usage{
hts_bin_telework_time(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified telework time variable
}
\description{
Add simplified telework time variable
Requires `telework_time` variable
}
\author{
Michael Jensen
}
