% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_transit_frequency}
\alias{hts_bin_transit_frequency}
\title{Add generalized transit frequency
Requires `transit_frequency` variable}
\usage{
hts_bin_transit_frequency(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a generalized transit frequency variable
}
\description{
Add generalized transit frequency
Requires `transit_frequency` variable
}
\author{
Michael Jensen
}
