% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-trip-vars.R
\name{hts_bin_transit_mode_acc}
\alias{hts_bin_transit_mode_acc}
\title{Add simplified transit access mode classification
Requires `mode_acc` variable}
\usage{
hts_bin_transit_mode_acc(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified transit access mode variable
}
\description{
Add simplified transit access mode classification
Requires `mode_acc` variable
}
\author{
Michael Jensen
}
