% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-hh-vars.R
\name{hts_bin_vehicle_count}
\alias{hts_bin_vehicle_count}
\title{Add simplified vehicle count variables
Requires `vehicle_count` variable}
\usage{
hts_bin_vehicle_count(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with simplified vehicle count variables
}
\description{
Add simplified vehicle count variables
Requires `vehicle_count` variable
}
\author{
Michael Jensen
}
