% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_walk_freq}
\alias{hts_bin_walk_freq}
\title{Add generalized walk frequency
Requires `walk_freq` variable}
\usage{
hts_bin_walk_freq(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a generalized walk frequency variable
}
\description{
Add generalized walk frequency
Requires `walk_freq` variable
}
\author{
Michael Jensen
}
