% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-person-vars.R
\name{hts_bin_walk_frequency}
\alias{hts_bin_walk_frequency}
\title{Add generalized walk frequency
Requires `walk_frequency` variable}
\usage{
hts_bin_walk_frequency(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a generalized walk frequency variable
}
\description{
Add generalized walk frequency
Requires `walk_frequency` variable
}
\author{
Michael Jensen
}
