% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-groupings.R
\name{hts_bin_worker}
\alias{hts_bin_worker}
\title{Add simplified worker status variable
Requires `employment` variable}
\usage{
hts_bin_worker(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified worker status variable
}
\description{
Add simplified worker status variable
Requires `employment` variable
}
\author{
Michael Jensen
}
