% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_hts.R
\name{hts_wgt_var}
\alias{hts_wgt_var}
\title{Return weight variable name}
\usage{
hts_wgt_var(tblname)
}
\arguments{
\item{tblname}{table name}
}
\value{
weight variable name
}
\description{
Return weight variable name
}
