% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_hts.R
\name{psrc_hts_recode_na}
\alias{psrc_hts_recode_na}
\title{Recode missing values to NA}
\usage{
psrc_hts_recode_na(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with missing values recoded NA
}
\description{
Recode missing values to NA
}
\author{
Michael Jensen
}
