% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_hts.R
\name{psrc_hts_to_factor}
\alias{psrc_hts_to_factor}
\title{Convert a column to factor datatype w/ levels specified from codebook}
\usage{
psrc_hts_to_factor(dt)
}
\arguments{
\item{dt}{the data.table}
}
\value{
the data.table with column in factor datatype
}
\description{
Convert a column to factor datatype w/ levels specified from codebook
}
\author{
Michael Jensen
}
