% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psrc_hts.R
\name{psrc_hts_varsearch}
\alias{psrc_hts_varsearch}
\title{Search PSRC travel survey variable definitions}
\usage{
psrc_hts_varsearch(regex)
}
\arguments{
\item{regex}{search term}
}
\value{
data.table of filtered variable attributes
}
\description{
Look for a variable name using a search term
}
\author{
Michael Jensen
}
