% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_hts_stat.R
\name{psrc_hts_vmtrate}
\alias{psrc_hts_vmtrate}
\title{Summarize PSRC travel survey vmt rates}
\usage{
psrc_hts_vmtrate(hts_data, group_vars = NULL, incl_na = TRUE)
}
\arguments{
\item{hts_data}{the data object, a list with either data.table or NULL for hh, person, day, trip, vehicle}

\item{group_vars}{vector with names of one or more grouping variables, in order}

\item{incl_na}{logical, whether NA should be included in results--including calculated shares}
}
\value{
summary table
}
\description{
Summarize PSRC travel survey vmt rates
}
\author{
Michael Jensen
}
