% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_sss_stat.R
\name{psrc_sss_stat}
\alias{psrc_sss_stat}
\title{Summarize PSRC transportation safety & security survey data}
\usage{
psrc_sss_stat(sss_data, group_vars = NULL, stat_var = NULL, incl_na = TRUE)
}
\arguments{
\item{sss_data}{data.table returned by `get_psrc_sss()`}

\item{group_vars}{vector with names of one or more grouping variables, in order}

\item{stat_var}{string, name of numeric variable for min/max/median/mean; implicit for count/share}

\item{incl_na}{logical, whether NA should be included in results--including calculated shares}
}
\value{
summary table
}
\description{
Summarize PSRC transportation safety & security survey data
}
\author{
Michael Jensen
}
