% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{z_score}
\alias{z_score}
\title{Z Score}
\usage{
z_score(x, y)
}
\arguments{
\item{x}{a row of aggregate table to compare--includes a count/sum/mean/median and its MOE}

\item{y}{a row of aggregate table to compare--also includes a count/sum/mean/median and its MOE}
}
\value{
Z score; if larger than 1, difference is significant
}
\description{
Stat to determine if two estimates are different
}
\author{
Michael Jensen
}
