% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{get_hhts}
\alias{get_hhts}
\title{HHTS data retrieval}
\usage{
get_hhts(survey, level, vars, ...)
}
\arguments{
\item{survey}{character string denoting single or combined survey year--if the latter, years separated by underscore, e.g."2017_2019"}

\item{level}{either "h" ("households"), "p" ("persons"), "d" ("days"), "t" ("trips"), or "v" ("vehicles")}

\item{vars}{character vector with requested variables}

\item{...}{pass-through for hhts parameters--either Elmer or SQLite}
}
\value{
dataframe with variables and necessary weights
}
\description{
Gets requested Household Travel Survey variables
}
