% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{get_var_defs}
\alias{get_var_defs}
\title{Retrieve HHTS variable definitions}
\usage{
get_var_defs(vars, ...)
}
\arguments{
\item{vars}{character vector with requested variables}

\item{...}{pass-through for hhts parameters--either Elmer or SQLite}
}
\value{
data.table of filtered variable attributes
}
\description{
Gets requested variable attributes--e.g. data type, associated weight name & priority
}
\author{
Michael Jensen
}
