% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{hhts2srvyr}
\alias{hhts2srvyr}
\title{HHTS to srvyr}
\usage{
hhts2srvyr(df, dyear, vars, spec_wgt = NULL, incl_na = TRUE)
}
\arguments{
\item{df}{the dataframe returned by \code{\link{get_hhts}}}

\item{dyear}{data year or data year vector, e.g. c(2017, 2019)}

\item{vars}{character vector with requested variables}

\item{spec_wgt}{optional user-specified expansion weight; only possible if the variable name is included in the \code{\link{get_hhts}} call.}

\item{incl_na}{option to remove NA from group_vars (if FALSE, the total will not reflect the full dataset)}
}
\value{
srvyr object with sampling weights
}
\description{
Creates srvyr object from HHTS
}
