% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{hhts2srvyr}
\alias{hhts2srvyr}
\title{HHTS to srvyr}
\usage{
hhts2srvyr(df, dyear, vars)
}
\arguments{
\item{df}{input dataframe}

\item{dyear}{data year or data year vector, e.g. c(2017, 2019)}

\item{vars}{character vector with requested variables}
}
\value{
srvyr object with sampling weights
}
\description{
Creates srvyr object from HHTS
}
