% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhts.R
\name{hhts_stat}
\alias{hhts_stat}
\alias{hhts_count}
\alias{hhts_sum}
\alias{hhts_median}
\alias{hhts_mean}
\alias{hhts_summary}
\title{Generic call for HHTS summary statistics}
\usage{
hhts_stat(df, stat_type, target_var, group_vars, geographic_unit = NULL)

hhts_count(df, target_var = NULL, group_vars = NULL, geographic_unit = NULL)

hhts_sum(df, target_var, group_vars = NULL, geographic_unit = NULL)

hhts_median(df, target_var, group_vars = NULL, geographic_unit = NULL)

hhts_mean(df, target_var, group_vars = NULL, geographic_unit = NULL)

hhts_summary(df, target_var, group_vars = NULL, geographic_unit = NULL)
}
\arguments{
\item{df}{the dataframe returned by \code{\link{get_hhts}}}

\item{stat_type}{Desired survey statistic}

\item{target_var}{The exact HHTS target variable intended}

\item{group_vars}{Factor variable/s for grouping}

\item{geographic_unit}{Default="region"}

\item{so}{The srvyr object}
}
\value{
A summary tibble, including variable names, summary statistic and margin of error

A table with the variable names and labels, summary statistic and margin of error
}
\description{
Given specific form by related \code{\link{hhts_stat}} functions.

Separate function for total, count, median, mean
}
