% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc-hts-recode-day-vars.R,
%   R/psrc-hts-recode-person-vars.R
\name{hts_bin_telecommute_freq}
\alias{hts_bin_telecommute_freq}
\title{Add simplified telework time variable
Requires `telecommute_freq` variable}
\usage{
hts_bin_telecommute_freq(hts_data)

hts_bin_telecommute_freq(hts_data)
}
\arguments{
\item{hts_data}{the hts_data list object}
}
\value{
hts_data with a simplified telework time variable

hts_data with a simplified telework time variable
}
\description{
Add simplified telework time variable
Requires `telecommute_freq` variable

Add simplified telework time variable
Requires `telecommute_freq` variable
}
\author{
Michael Jensen
}
