% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_hts_stat.R
\name{psrc_hts_stat}
\alias{psrc_hts_stat}
\title{Summarize PSRC travel survey data}
\usage{
psrc_hts_stat(
  hts_data,
  analysis_unit,
  group_vars = NULL,
  stat_var = NULL,
  incl_na = TRUE
)
}
\arguments{
\item{hts_data}{the data object, a list with either data.table or NULL for hh, person, day, trip, vehicle}

\item{analysis_unit}{string, either "household", "person", "day", "trip", or "vehicle"}

\item{group_vars}{vector with names of one or more grouping variables, in order}

\item{stat_var}{string, name of numeric variable for sum, median, mean; implicit for count}

\item{incl_na}{logical, whether NA should be included in results--including calculated shares}
}
\value{
summary table
}
\description{
Summarize PSRC travel survey data
}
\author{
Michael Jensen
}
